generator/pyproject.toml
[tool.poetry]
name = "generator"
version = "0.1.0"
description = "Generator for Leonidas and Cloud Detection Docs"
authors = ["Nick Jones"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
jinja2 = "^2.11.2"
mkdocs = "^1.1.2"
mkdocs-material = "^5.2.1"
pyyaml = "^5.3.1"
black = "^19.10b0"
pylint = "^2.5.2"
typer = "^0.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
